Create docker-compose.yaml file
Use docker-compose for container orchestration to make a postgres
database available for usage with the chatbot Data API app. This
database is necessary to store the user inputs after receiving
their consent. This way, data scientist have access to the inputs
saved in the persistent storage later to further improve the
chatbot.

Signed-off-by: Iuliana Voinea <iuliana.voinea@swisscom.com>